Added check_syntax3
