+++ Makefile
@@ -6,7 +6,7 @@
 #    By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/09/03 22:33:20 by irychkov          #+#    #+#              #
-#    Updated: 2024/10/22 12:15:45 by irychkov         ###   ########.fr        #
+#    Updated: 2024/10/22 13:31:09 by irychkov         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -15,7 +15,7 @@ NAME = minishell
 SRCS = src/main.c src/split_args.c src/split_args_helper.c \
 		src/init_pipes.c src/init_cmd.c src/here_doc.c \
 		src/ft_update_args.c src/parse_redirect.c \
-		src/check_syntax.c src/check_syntax2.c \
+		src/check_syntax.c src/check_syntax2.c src/check_syntax3.c\
 		src/trim_spaces.c \
 		src/exec_cmd.c src/exec_pipes.c \
 		src/errors.c src/sys_errors.c\
